fix: move simulator to a specific file
